Focus behavior of the URL entry field
Opening a new tab will now set input focus to the URL field.

Switching tabs when the URL field is focused will select the full new URL.

There was an issue previously where selecting a bookmark while the URL field was focused would cause unexpected behavior. The URL was first unfocused and that would trigger the document to fetch the URL that was in the input field.