Fix opening of mailto: links
This code path is used for opening both http/s and mailto links. On
macOS, open is used generically handling both nicely. On Linux,
x-www-browser was used for both, meaning it would attempt to open mailto
links with the browser. This -does- work but is clunky. On my system, my
defaults are Firefox for web and mutt for email, so if I clicked a
mailto, Firefox would open (slow!) just to go out of focus as it opens
mutt.

Let's just mirror the macOS behaviour and use xdg-open for both kinds of
links. Then the browser will be used for http/s and the mail client for
mailto and everyone's happy.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>