GmResponse: Fixed parsing of the response
In some cases an empty <META> in the server response would cause GmResponse to treat the entire header as invalid.

Also, TlsRequest was not handling closing of the TLS/SSL connection if the socket still remained open. This caused GmRequest to wait until the timeout before it realized the full response had been received.

IssueID #22