ui: Fix line wrapping of emoticons
While we do want to have special handling of punctuation to help wrap
pathologically-long-hyphenated-words-that-go-on-for-kilometers, we don't
want to wrap emoticons, so let's adjust the logic to account for this.
While we're at it, clean up and add a comment explaining the logic for
the benefit of future readers.

Fixes rendering of a recent post on CAPCOM at certain window sizes

=> gemini://rosenzweig.io/lagrange/Before.png Before the change
=> gemini://rosenzweig.io/lagrange/After.png After the change

Notice the wrapping of the ":D" in the second paragraph.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>