hoverLink: clear on mouse clicks
When selecting a link item with the mouse, if one had done so with the
right-mouse button context meny, then the highlighting of that link
would not have been cleared when scrolling the page, or hovering over
another link.

When considering which links are hover links, unconditionally
dehighlight all links before highlighting the link under the mouse
cursor.