Windows: Custom frame behavior
A borderless SDL window gets no standard window behavior, so
this commit implements support for some of them. There is
special handling for the window frame and various snapping modes.

Not quite finished yet... It might make sense to research if
a custom window class could work with this app (with SDL); could be
less work.