Single app instance; IPC mechanism
Only one instance of Lagrange is allowed to run per a runtime directory. Otherwise the instances will overwrite each others' files.

Added a check at startup to see if an instance is already running. If so, options can be used to communicate with it.

By default, a new tab is opened in the running instance.

IssueID #37
IssueID #164
IssueID #174
IssueID #178