app: Raise windows when opening new tabs via IPC
When Lagrange is opened via a terminal -- or more importantly, when
opened as a MIME handler for gemini:// links in other applications --
users expect Lagrange to take focus. However, if Lagrange is already
open and the request is piped over IPC, this fails. In theory this
should be as simple as adding a SDL_RaiseWindow() in the right place. In
practice there's a chain of race conditions that make the obvious
solution fail in a non-obvious way.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>