File URLs support directories and ZIP archives
One is now able to view directory and ZIP archive contents using file URLs.

Directory contents are shown as a list of links, enabling previewing supported formats. A link to the parent directory is included at the top of the page.

ZIP archives behave like directories, except they also recognize "index.gmi" and "index.gemini" files and display them instead of the regular directory index (when viewing a directory inside a ZIP archive). This enables archiving a Gemini capsule and browsing it as a ZIP archive.

Added a preference to disable loading of index pages.