DocumentWidget: Fixed drawing of empty document
An empty document would could the widget to be drawn as a black non-themed rectangle, with a single-pixel horizontal line in the middle. This fixes that by just clearing the widget to the UI background color.

The most notable situation where this occurred was immediately after activating split view mode.