ListWidget: Dragging before/after an item
It's useful to make a distinction whether a drag ended before or after an item, even if it's still referring to the same boundary between items.

This allows bookmarks to be reordered inside a folder so that an item is moved to the bottom of a folder, or out of the folder following it in order.