Media refactoring; working on FontPack management
Media still needs more work to get rid of redundancies and make lookups faster.

FontPacks are manipulated as Media items (not unlike images) so they can be previewed on page, and installed via a click.

FontPack management is not trivial as it includes such details as versioning and whether individual packs are enabled or disabled.