Text attributes that change inside a run
These changes concern the situation when the attributes of text (i.e., font, color) are changed via escape sequences.

The concept of "base attributes" was added so that the low-level text renderer knows which font/color to set when a "reset" escape sequence is encountered. This depends on what kind of text is being renderer, e.g., preformatted or regular paragraphs.

The base attributes were added as variables in Text because it was getting unwieldy to pass all the information via the draw/measure/WrapText functions.

GmDocument now has a GmTheme struct that collects the font and color information into a single place.