GmDocument: Remove heading line overlap
Spacing of heading lines was being reduced for a nicer appearance, but since backgrounds are erased behind text runs, it was causing issues with some fonts.

Therefore this is disabled for now until I figure out how I want to fix the background issue. Maybe smarter clipping could be applied in the page renderer so that runs are only drawn once, and erasing the background becomes unnecessary?

IssueID #364