Feeds: Fixed handlig of duplicate entries
Feed entries are expected to have unique URLs, since the URL is the identifier of an entry. If a feed has the same URL for more than one entry, only process the first one and ignore the rest.

An entry will still show up as unread if it keeps the same URL but changes its title.