Image metadata; improved open URLs indicator
Put a new GmRun under an inlined image with the image metadata, so it doesn't overlap the caption text.

Open URLs are now indicated with a document-width box instead of window-width box. The box is also slightly expanded for a nicer look. Links have a bit of padding around them so this doesn't cause overlaps with other runs.