Handling keys while navigating via home row
When navigating via home row, ensure that other widgets will not trigger when keys are pressed. The document is given keyboard focus even though it is not flagged as focusable, to make sure it gets to handle all the keys first.

On macOS, disable all native menu items that have a home row keyboard shortcut.

Also on macOS, disable all native menus when changing bindings.

IssueID #419