Fonts: Improved fallback glyph search
The fontpack documentation says that when a glyph isn't found, all available fonts are checked in priority order. However, the implementation didn't actually do this.

Now there is a separate priority mapping for loaded fonts so they can be quickly scanned for any missing glyphs in the right order.

"iosevka-body" was prioritized higher because it provides a number of UI icons.