Placing navbar and tab bars at the bottom
There are now two separate preferences for moving the navbar and/or the tab bar to the bottom.

These required quite extensive animation refinements for the phone layout. The phone toolbar now has a parent element (bottom bar) that can also optionally house the navbar. The tab bar is never moved outside the tabs widgets, but is still animated with an offset to stay above the bottom bar.