Opening multiple bookmarks at once in a sequence
When opening a folder of bookmarks, it's not the best idea to start
multiple concurrent network requests. Something like the controlled
queue that the feeds refresher uses would be nice, but in the
meantime, documents just wait until the previous one is finished
before starting to fetch a new URL.

This behavior is only used when opening a folder of bookmarks.