Improved tab-specific pinned identities
The pinned identity of a tab is saved persistently to state.lgr.
When opening links in the tab, always open them in new tabs if the
destination is expected to have a different identity.

Improved handling of the bookmark identity property throughout
the app, for example when clicking on the "bookmarked" indicator in
the navbar.