Mobile: Fixed swipe navigation and scroll position issues
Avoid flicker during swipe navigation, and try to retain the
scroll position more closely.

The swipe animation is done with the help of additional
DocumentWidgets, but that is prone to various side effects, such as
overwriting the scroll position in the history, and rendering a
frame or two of invalid content. (This technical debt really should
be paid sooner or later.)