DocumentWidget: Revising implementation of swipe animations
Instead of multiple widgets, do swipe animation with the existing
DocumentView that was just being shown normally.

The state of the swipe is tracked with a number of new internal
flags in DocumentWidget. Take note of the handling of aborted
swipes, where additional flags are used to indicate that the swipe
animation should not be reset at finish but kept at the final
swipe position until the new page is ready.