Remember user-resized widget widths
Introduced a new mechanism for storing widget widths based on a
new "resize ID" string. The resize IDs are not meant to be unique,
so any bookmark editor, upload dialog, or input prompt can reuse
the previously resized widths.

This might be useful for sidebars as well, however those are saved
individually because they are saved and restored as part of the
"state.lgr" on a per-window/per-widget basis.