Widget: Removed keyboard focus kludge
When opening a query prompt through link number mode and holding
the modifier key, the opened prompt would not have input focus.
There was a special override giving the DocumentWidget keyboard
focus, taking it away from the InputWidget without the latter
being notified.

This kludge was ill-advised as it can break normal focus logic.
It likely became an issue when the behavior of holding the modifier
key was fixed.